s3_reproduceability/final_exercise/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
seaborn
matplotlib
scikit-learn
numpy
hydra-core
omegaconf
torch==1.10.1
torchvision==0.11.2
torchaudio==0.10.1